requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0,<4.0
django-cors-headers>=3.13.0,<4.0
django-filter>=23.1.0,<24.0
djangorestframework-simplejwt>=5.2.0,<6.0
psycopg2-binary>=2.9.6,<3.0